requirements.txt
pandas
requests
beautifulsoup4
pytz
anvil-uplink
gspread
oauth2client
